Separated atoms and organisms in scss/ directory
