+++ client/src/dev-components/container-card/container-card.scss
@@ -1,7 +1,7 @@
 @import '../../scss/compulsory';
 
-@import '../../scss/atoms-organisms/button';
-@import '../../scss/atoms-organisms/tooltip';
+@import '../../scss/atoms/button';
+@import '../../scss/atoms/tooltip';
 
 .instance-card {
   flex: 1;
