+++ client/src/dev-components/create-runner/create-runner.scss
@@ -2,7 +2,7 @@
 @import '../../scss/colors';
 @import '../../scss/mixins';
 
-@import '../../scss/atoms-organisms/tooltip';
+@import '../../scss/atoms/tooltip';
 @import '../create/create';
 
 .runner-view-container {
