+++ client/src/dev-components/create/create.scss
@@ -1,7 +1,7 @@
 @import '../../scss/compulsory';
 
-@import '../../scss/atoms-organisms/button';
-@import '../../scss/atoms-organisms/forms';
+@import '../../scss/atoms/button';
+@import '../../scss/organisms/forms';
 
 @import '../dashboard-view/dashboard-view';
 @import '../../feedbacker-components/reactions/reactions';
