+++ client/src/dev-components/dashboard-view/dashboard-view.scss
@@ -1,7 +1,7 @@
 @import '../../scss/compulsory';
 
-@import '../../scss/atoms-organisms/button';
-@import '../../scss/atoms-organisms/tooltip';
+@import '../../scss/atoms/button';
+@import '../../scss/atoms/tooltip';
 
 .top-section {
   display: flex;
