+++ client/src/feedbacker-components/add-username-modal/add-username-modal.scss
@@ -2,9 +2,9 @@
 @import '../../scss/colors';
 @import '../../scss/mixins';
 
-@import '../../scss/atoms-organisms/forms';
-@import '../../scss/atoms-organisms/button';
-@import '../../scss/atoms-organisms/modal';
+@import '../../scss/organisms/forms';
+@import '../../scss/atoms/button';
+@import '../../scss/atoms/modal';
 
 @import '../comment/comment.scss';
