+++ client/src/feedbacker-components/comment-label/comment-label.scss
@@ -1,7 +1,7 @@
 @import '../../scss/variables';
 @import '../../scss/colors';
 @import '../../scss/mixins';
-@import '../../scss/atoms-organisms/button';
+@import '../../scss/atoms/button';
 
 
 .developer-label {
