+++ client/src/feedbacker-components/comment-panel/comment-panel.scss
@@ -1,7 +1,7 @@
 @import '../../scss/compulsory';
 
-@import '../../scss/atoms-organisms/button';
-@import '../../scss/atoms-organisms/forms';
+@import '../../scss/atoms/button';
+@import '../../scss/organisms/forms';
 
 @import '../survey-panel/survey-panel.scss';
 @import '../comment/comment.scss';
