+++ client/src/feedbacker-components/confirm-modal/confirm-modal.scss
@@ -1,7 +1,7 @@
 @import '../../scss/variables';
 @import '../../scss/colors';
 @import '../../scss/mixins';
-@import '../../scss/atoms-organisms/button';
+@import '../../scss/atoms/button';
 
 @import '../comment/comment.scss';
