+++ client/src/feedbacker-components/open-survey-panel-button/open-survey-panel-button.scss
@@ -1,6 +1,7 @@
 @import '../../scss/compulsory';
+
+@import '../../scss/atoms/button';
 @import '../../scss/animations';
-@import '../../scss/atoms-organisms/button';
 
 @import '../open-comment-panel-button/open-comment-panel-button.scss';
