+++ client/src/feedbacker-components/reactions/reactions.scss
@@ -1,6 +1,6 @@
 @import '../../scss/compulsory';
 
-@import '../../scss/atoms-organisms/button';
+@import '../../scss/atoms/button';
 
 .reactions {
   display: flex;
