+++ client/src/feedbacker-components/route-container/route-container.scss
@@ -3,8 +3,8 @@
 @import '../../scss/colors';
 @import '../../scss/mixins';
 
-@import '../../scss/atoms-organisms/button';
-@import '../../scss/atoms-organisms/anchor';
+@import '../../scss/atoms/button';
+@import '../../scss/atoms/anchor';
 
 .route-container {
   position: relative;
