+++ client/src/feedbacker-components/submit-field/submit-field.scss
@@ -1,6 +1,6 @@
 @import '../../scss/compulsory.scss';
-@import '../../scss/atoms-organisms/button';
-@import '../../scss/atoms-organisms/checkbox';
+@import '../../scss/atoms/button';
+@import '../../scss/atoms/checkbox';
 
 @import '../survey-panel/survey-panel.scss';
 @import '../comment/comment.scss';
