+++ client/src/feedbacker-components/survey-edit/survey-edit-container.scss
@@ -2,7 +2,7 @@
 @import '../../scss/colors';
 @import '../../scss/mixins';
 
-@import '../../scss/atoms-organisms/button';
+@import '../../scss/atoms/button';
 
 .drag-helper {
   @include z-level(1);
