+++ client/src/feedbacker-components/survey-panel/survey-panel.scss
@@ -1,6 +1,6 @@
 @import '../../scss/compulsory';
 
-@import '../../scss/atoms-organisms/button';
+@import '../../scss/atoms/button';
 
 .panel-container {
   @include display-and-animate();
