+++ client/src/feedbacker-components/survey/survey.scss
@@ -1,8 +1,8 @@
 @import '../../scss/compulsory';
 // Atoms
-@import '../../scss/atoms-organisms/button';
-@import '../../scss/atoms-organisms/forms';
-@import '../../scss/atoms-organisms/progress-bar';
+@import '../../scss/atoms/button';
+@import '../../scss/organisms/forms';
+@import '../../scss/atoms/progress-bar';
 
 @import '../reactions/reactions';
