+++ client/src/feedbacker-components/tag-element-button/tag-element-button.scss
@@ -1,6 +1,6 @@
 @import '../../scss/compulsory';
 
-@import '../../scss/atoms-organisms/button';
+@import '../../scss/atoms/button';
 
 @import '../comment/comment';
