+++ client/src/index.js
@@ -25,7 +25,7 @@ import { setupPersist } from './persist'
 import { getSession, updateSession } from './session'
 import { loadPersistData, setPersistData, loadComments, updateRole } from './actions'
 // Styles
-import './scss/atoms-organisms/_toast.scss'
+import './scss/atoms/_toast.scss'
 import styles from './scss/_base.scss'
 
 const css = classNames.bind(styles)
