+++ client/src/scss/atoms/_button.scss
@@ -1,4 +1,4 @@
-@import '../../scss/compulsory';
+@import '../compulsory';
 
 %accent {
   box-shadow: 0 2px 5px $palette-shadow-default;
