+++ client/src/scss/atoms/_checkbox.scss
@@ -1,3 +1,5 @@
+@import '../compulsory';
+
 $checkbox-width: $baseline * 3;
 
 input[type="checkbox"], %checkbox {
