+++ client/src/scss/atoms/_modal.scss
@@ -1,7 +1,7 @@
-@import '../../scss/variables';
-@import '../../scss/colors';
-@import '../../scss/mixins';
-@import '../../scss/atoms-organisms/button';
+@import '../variables';
+@import '../colors';
+@import '../mixins';
+@import 'button';
 
 $padding: $spacer * 2;
