+++ client/src/scss/atoms/_progress-bar.scss
@@ -1,4 +1,4 @@
-@import '../../scss/compulsory';
+@import '../compulsory';
 
 .progress-bar {
   position: relative;
