+++ client/src/scss/atoms/_tooltip.scss
@@ -1,6 +1,4 @@
-// TODO: change variables to compulsory
-@import '../variables';
-@import '../typography';
+@import '../compulsory';
 
 [data-tooltip] {
   position: relative;
