+++ client/src/scss/organisms/_forms.scss
@@ -1,6 +1,6 @@
-@import '../../scss/compulsory';
+@import '../compulsory';
 
-@import '../../scss/atoms-organisms/button';
+@import '../atoms/button';
 
 %textarea {
   padding: $spacer;
